Created `resource_management`, fixed `layer_collection` visibility issues
